requirements.txt
Django==5.0.4
djangorestframework==3.15.2
psycopg2==2.9.10
python-decouple==3.8
pytz==2024.1
tzdata==2024.1
sqlparse==0.5.0
django-allauth
requests
cryptography
Pillow>=9.0.0
django-cors-headers
gunicorn
whitenoise
djangorestframework-simplejwt